:rocket: Register errors in app initialization
