+++ labsys/main/errors.py
@@ -1,17 +1,1 @@
-from flask import render_template
 
-from .views import blueprint
-
-@blueprint.app_errorhandler(403)
-def page_not_found(e):
-    return render_template('403.html'), 403
-
-
-@blueprint.app_errorhandler(404)
-def page_not_found(e):
-    return render_template('404.html'), 404
-
-
-@blueprint.app_errorhandler(500)
-def internal_server_error(e):
-    return render_template('500.html'), 500
