+++ labsys/templates/403.html
@@ -1,6 +1,6 @@
 {% extends "base.html" %}
 
-{% block title %}LabSys- Acesso Negado{% endblock %}
+{% block title %}LabSys - Acesso Negado{% endblock %}
 
 {% block page_content %}
 <div class="page-header">
