+++ labsys/templates/404.html
@@ -1,9 +1,9 @@
 {% extends "base.html" %}
 
-{% block title %}LabSys - Page Not Found{% endblock %}
+{% block title %}LabSys - Página não encontrada{% endblock %}
 
 {% block page_content %}
 <div class="page-header">
-    <h1>Not Found</h1>
+    <h1>Não encontrada</h1>
 </div>
 {% endblock %}
